Implementation of mvc architecture
