+++ asset/Src/Views/index.html
@@ -3,12 +3,12 @@
   <head>
     <meta charset="UTF-8" />
     <title>Chat</title>
-    <link rel="stylesheet" href="./asset/style/style_chat.css" />
-    <link rel="stylesheet" href="./asset/style/style_chat_folders.css" />
-    <link rel="stylesheet" href="./asset/style/style_chat_chatlist.css" />
-    <link rel="stylesheet" href="./asset/style/style_chat_dialog.css" />
-    <link rel="stylesheet" href="./asset/style/style_icons_and_fonts.css" />
-    <link rel="stylesheet" href="./asset/style/style_add_contacts.css" />
+    <link rel="stylesheet" href="../../style/style_chat.css" />
+    <link rel="stylesheet" href="../../style/style_chat_folders.css" />
+    <link rel="stylesheet" href="../../style/style_chat_chatlist.css" />
+    <link rel="stylesheet" href="../../style/style_chat_dialog.css" />
+    <link rel="stylesheet" href="../../style/style_icons_and_fonts.css" />
+    <link rel="stylesheet" href="../../style/style_add_contacts.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
     <script
@@ -78,7 +78,7 @@
           <div class="dialog__header">
             <div class="dialog__header-right">
               <img
-                src="./asset/image/user.png"
+                src="../../image/user.png"
                 alt="profile"
                 class="message__img"
               />
@@ -103,7 +103,7 @@
             <div class="message message__self">
               <div class="message__photo">
                 <img
-                  src="./asset/image/user.png"
+                  src="../../image/user.png"
                   alt="profile"
                   class="message__img"
                 />
@@ -259,7 +259,6 @@
       integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
       crossorigin="anonymous"
     ></script>
-    <script src="./asset/js/pageReaction.js"></script>
-    <script src="./asset/js/voice.js"></script>
+    <script src="../../js/pageReaction.js"></script>
   </body>
 </html>
