+++ asset/Src/Views/login.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <title>Sign In & Sign Up</title>
-    <link rel="stylesheet" type="text/css" href="./asset/style/style_login.css" />
+    <link rel="stylesheet" type="text/css" href="../../style/style_login.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
   </head>
 
@@ -13,7 +13,7 @@
         onchange="ValidationSignup()"
         class="form-main sign-up-container"
       >
-        <form name="Signup" id="Signup" action="./chat.html" method="post">
+        <form name="Signup" id="Signup" action="./index.html" method="post">
           <h1>ساخت حساب کاربری</h1>
           <input
             class="input"
@@ -65,7 +65,7 @@
         onchange="ValidationSignin()"
         class="form-main sign-in-container"
       >
-        <form name="Signin" id="Signin" action="./chat.html" method="post">
+        <form name="Signin" id="Signin" action="./index.html" method="post">
           <h1>ورود به حساب</h1>
           <input
             class="input"
@@ -121,6 +121,6 @@
         </div>
       </div>
     </div>
-    <script src="./asset/js/validationSignin.js"></script>
+    <script src="../../js/validationSignin.js"></script>
   </body>
 </html>
