+++ asset/Src/db/init.php
@@ -1,7 +1,5 @@
 <?php
 require 'rb-mysql.php';
-
-try{
     // R::setup(
     //     'mysql:host=denali.liara.cloud;port=30653;dbname=nostalgic_dhawan',
     //     'root',
@@ -12,7 +10,3 @@ try{
         'root',
         ''
     );
-    
-}catch(Exception $err){
-    die('connection failed: '.$err->getMessage());
-}
