--- asset/php/index.php
@@ -1,3 +1,0 @@
-<?php
-
-include 'init.php';
