--- asset/php/models/messages/fetch.php
@@ -1,11 +1,0 @@
-<?php
-
-require 'message.php';
-
-$message=new Message();
-$data = $message->selectAllData();
-R::close();
-
-echo $data;
-
-exit();
